Add tier & tag source
