+++ openmetadata-service/src/main/resources/elasticsearch/en/api_collection_index_mapping.json
@@ -423,6 +423,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
