+++ openmetadata-service/src/main/resources/elasticsearch/en/api_endpoint_index_mapping.json
@@ -648,6 +648,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
