+++ openmetadata-service/src/main/resources/elasticsearch/en/classification_index_mapping.json
@@ -127,6 +127,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "domain" : {
         "properties": {
           "id": {
