+++ openmetadata-service/src/main/resources/elasticsearch/en/container_index_mapping.json
@@ -588,6 +588,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
