+++ openmetadata-service/src/main/resources/elasticsearch/en/dashboard_data_model_index_mapping.json
@@ -451,6 +451,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
