+++ openmetadata-service/src/main/resources/elasticsearch/en/dashboard_index_mapping.json
@@ -527,6 +527,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
