+++ openmetadata-service/src/main/resources/elasticsearch/en/database_index_mapping.json
@@ -498,6 +498,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
