+++ openmetadata-service/src/main/resources/elasticsearch/en/database_schema_index_mapping.json
@@ -463,6 +463,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
