+++ openmetadata-service/src/main/resources/elasticsearch/en/domain_index_mapping.json
@@ -278,6 +278,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "extension": {
         "type": "object"
       }
