+++ openmetadata-service/src/main/resources/elasticsearch/en/glossary_term_index_mapping.json
@@ -449,6 +449,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "childrenCount": {
         "type": "integer"
       }
