+++ openmetadata-service/src/main/resources/elasticsearch/en/ingestion_pipeline_index_mapping.json
@@ -350,6 +350,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "applicationType": {
         "type": "keyword",
         "normalizer": "lowercase_normalizer"
