+++ openmetadata-service/src/main/resources/elasticsearch/en/metric_index_mapping.json
@@ -348,6 +348,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
