+++ openmetadata-service/src/main/resources/elasticsearch/en/mlmodel_index_mapping.json
@@ -524,6 +524,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
