+++ openmetadata-service/src/main/resources/elasticsearch/en/pipeline_index_mapping.json
@@ -438,6 +438,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
