+++ openmetadata-service/src/main/resources/elasticsearch/en/search_entity_index_mapping.json
@@ -480,6 +480,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
