+++ openmetadata-service/src/main/resources/elasticsearch/en/stored_procedure_index_mapping.json
@@ -497,6 +497,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
