+++ openmetadata-service/src/main/resources/elasticsearch/en/table_index_mapping.json
@@ -653,6 +653,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "columnDescriptionStatus": {
         "type": "keyword"
       },
