+++ openmetadata-service/src/main/resources/elasticsearch/en/topic_index_mapping.json
@@ -514,6 +514,12 @@
       "descriptionSources": {
         "type": "object"
       },
+      "tagSources": {
+        "type": "object"
+      },
+      "tierSources": {
+        "type": "object"
+      },
       "certification": {
         "type": "object",
         "properties": {
